producer/requirements.txt
pandas
pyarrow
kafka-python
boto3

